Finish Integration With Relay
Added a system called the Timeline which allows work to be
set to run at specific points in time. This allows work to
be queued.

To allow the relay server data to merge in with the server,
a work item is set to run every minute. When it finishs, it
sets itself to run again in a minute.

When data needs to be sent to the relay, a work item is added
to the timeline so that the current connection can be closed
before sending data to the relay.

The timeline allowed for the Hub to be removed as work items
to handle connections could be added to the timeline.

Closes #4.