Made Uuid a class
Before Uuid was an interface, but every implementation was
the same. Every time someone made implemented Uuid, they
would have to call Uuids.complete on it to fill in the
missing functions.

Making Uuid a class solves all these problems as all the
methods will be implemented correctly and there only needs
to be one implementation.

Generators are still sepearate.