Exposing params in run scripts
Before the run script had defaults for most values. This
starting causing problems when hosting the server as it
would require the run scripts to change.

Now all the paramters are passed in when calling the run
script. If any value is missing, the script exits and a
usage message is printed.