Relay re-load teams file
Before the relay would only read the teams file at the start-up. This
meant that the relay would have to be restarted whenever a new team
was added.

This change makes the relay reload the teams file every minute so that
if a new team is added, they will have access within a minute. This will
prevent the relay having to be taken down and losing all its data each
time a new team comes on line.