Refining Client
Refining the commands that the text client uses by creating a simplier
context-based system for working with the server.

When reading the name of a user or converasion, read the whole line so that
we avoid names and titles from being cut-off.

The GUI client was removed to simplier for change.