Fixing Uuid Parse Error
In Uuid, ids are 32 bit unsigned integers. As Java does not supported
unsigned integers, when converting to a string, the 32 bit integer is
converted to a long so that it can be written to a string as a
positive integer (using all 32 bits).

When converting from a string to a Uuid, the integer value is not
handled correctly as values that would be acceptable as a 32 bit
unsigned integer fail to be parsed as a 32 bit signed integer.

If the id value of a Uuid is 0xFFFFFFFF and is converted to the string
4294967295. When 4294967295 is parsed, it will fail as it is not
expressible as a 32 bit signed integer.

Parsing needs to parse the unsigned integer and convert and store the
bit pattern correctly in a signed integer.

Closes #62