Client refactoring (#64)
Refining Client

Moving common files for clients into 'client.core' so that it is clear which files make up the shared core of all clients. The GUI client was removed to simpler for change.

Refining the commands that the text client uses by creating a simpler context-based system for working with the server. When reading the name of a user or conversion, read the whole line so that we avoid names and titles from being cut-off.  

To help students understand what the code is doing and what it is used for, this change adds more comments to explain what is going on and why.

Scanner is a large utility. It was being used to read a single line. Changed over to BufferedReader which is a lighter-weight way to read a line.