Server conversation refactoring (#67)
Clean Up Server Communication

To make things easier, the conversation was split into two parts: header and payload. This will make it easier to update the part that can be updated while not having to update the part that does not update.

Before server commands were handles as one long if-else chain. Now commands are defined as objects allowing them to be more isolated. Commands are kept in a map allowing them to easily be found given the command type.