Merge summer into master (#3)
* Made Secret into a full class

Before Secret only housed the parse function which would return a
byte array. Working with secrets would require using the Arrays
class. Now all the operations have been moved into Secret.

* Simplifying Secret Parse

Instead of adjusting the read index, it is easier to prepend
the string with "0" if the length is odd.

* Client refactoring (#64)

Refining Client

Moving common files for clients into 'client.core' so that it is clear which files make up the shared core of all clients. The GUI client was removed to simpler for change.

Refining the commands that the text client uses by creating a simpler context-based system for working with the server. When reading the name of a user or conversion, read the whole line so that we avoid names and titles from being cut-off.  

To help students understand what the code is doing and what it is used for, this change adds more comments to explain what is going on and why.

Scanner is a large utility. It was being used to read a single line. Changed over to BufferedReader which is a lighter-weight way to read a line.

* Server conversation refactoring (#67)

Clean Up Server Communication

To make things easier, the conversation was split into two parts: header and payload. This will make it easier to update the part that can be updated while not having to update the part that does not update.

Before server commands were handles as one long if-else chain. Now commands are defined as objects allowing them to be more isolated. Commands are kept in a map allowing them to easily be found given the command type.