Tool for accessing logs of CodeU-hosted servers.
This is a python script that alls one to access the logs of CodeU-hosted
servers. It can access a team's server's log or the relay log. It can
either pull the entire log or "watch" it (i.e., like "tail -f").

See the script source for information about using it.