Fixed issue with leaving conversations
Fixed an issue where the user would not properly leave an old
conversation, after they made a new conversation.

This was caused by the fact that when a client added a conversation,
they would automatically set their current conversation summary to that
of the new conversation, without notifying the server that they were
switching conversations. This resulted in them not properly leaving
their old conversation and therefore getting new messages for both.