Move Whole Project Into IntelliJ
Before we tried avoiding any IDEs, but this made it harder for others
to pick-up the project since it depended on many one-off build
scripts.

Using IntelliJ may make it harder to set up CI, but considering how
common IntelliJ is as a platform, there is likely a solution for
that. Regardless, since mass development of this project has dropped,
manual verification should be sufficient for now.