Simplify Logging Structure
Before we were wrapping the java logging library. This added a large
layer of complexity to the system. In this change we remove the use
of the java logging library and do a simple one-off implementation.

We omit the "write to file" functionality as it would make a good
initial feature of students to implement.