Remove Custom UUID Class
Originally the UUID class was used to allow for UUID chains. For
example a message's id would have a conversation's id as a parent
UUID in the UUID chain.

This feature of the custom UUID class we never actually used in any
meaningful way. This change removes the custom class and uses the
standard Java UUID class.