Address Analysis Warning About Threading
There was a busy wait in the time line logic. The busy wait was for
a feature that was not being used. By removing that feature, the code
became much simpler. Adding this feature back later, if needed, is
better than leaving the current implementation since a busy wait could
be avoided with some additional work.