Simplify Model
The overall model was complicated. By simplifying it, we were able
to simplify the views and controllers.

Continuing with this, the number of network commands needed went
down. Additional user commands were removed as they were better
left to be implemented by students. The purpose of the project is
to provide the skeleton.

By reducing the overall complexity, we should be able to improve
the test coverage and provide a better example of what testing
would look like in a larger project.