Move All Server Request-Response Logic Into Strategies
Before we had all the request-response logic in one large function.
The original plan with this was to write some "bad" code and use it
in a refactoring exercise. However, we found that refactoring
exercises we an ineffective tool.

Since most students wanted to add features, making it easier for them
to extend the current implementation is more important. By using
a strategies pattern right away, we can make the code easier to read,
maintain, and extend.