Remove Custom Time Type
Originally we had a wrapper for Date to make it more accessible, but
overall, the wrapper provided us with less value than we originally
wanted from it.

This change removes the wrapper and makes it so that we are using Date
directly. The largest part of the change is to the serialization code.